+++ noodles-csi/src/io/indexed_records/record.rs
@@ -5,9 +5,7 @@ use std::{error, fmt, ops::Range};
 use noodles_core::Position;
 
 use self::position::parse_start_position;
-use crate::{
-    binning_index::index::header::format::coordinate_system::CoordinateSystem, io::IndexedRecord,
-};
+use crate::{binning_index::index::header::format::CoordinateSystem, io::IndexedRecord};
 
 /// An indexed record.
 pub struct Record {